refactor(backend): reduce learning-store entrypoint surface
